Refactored helpers.py -> instantiated.py
